[9810] fbdev: dpu: allocate test framebuffer.
If CONFIG_FB_TEST is enabled, test framebuffer is allocated
for FB test.

Originally, framebuffer is allocated but it is freed when
first S3CFB_WIN_CONFIG ioctl is called.

Of course, framebuffer allocated in DECON probe is remained
in special booting mode using FB interface only not S3CFB_WIN_CONFIG.

This patch allocates another framebuffer for testing FB interface
in android booting mode.

It is only for test. So, CONFIG_FB_TEST is disabled as default

Change-Id: Ie06c182f258618d0dd212e9d9e3e6442dea23ffd
Signed-off-by: Jiun Yu <jiun.yu@samsung.com>