[9810] fbdev: dpu: DPU_2_0_RELEASE_FENCES feature is added
Previously, the display driver created retire_fence and passed it to
the HWComposer. The HWComposer then copies the retire_fence fd to the
number of layers it uses and passes it to surfaceflinger.

However, this patch was made to bring the display driver from the
HWComposer to create the release fence fd.

If DPU_2_0_RELEASE_FENCES is enables, the display driver creates
release_fence fd as many as the number of enabled windows and passes
it to HWComposer.

Change-Id: I0b9f1885666970f032fedbca7ce56ffb4e030f80
Signed-off-by: hwangjae lee <hj-yo.lee@samsung.com>