[9810] fbdev: dpu: the priority of the update thread has changed
The priority of decon_update_thread has been increased to real time
task. Because the priority of the other task in the anroid-O version is
set higher than decon_update_thread, and the operation of
decon_update_thread is delayed.

The problem of delaying the operation of decon_update_thread after the
priority change has been solved.

Change-Id: I77ec52adc03f677dd23f0a8107ce9156a1839f40
Signed-off-by: hwangjae lee <hj-yo.lee@samsung.com>