fbdev: dpu20: update BTS to consider initial latency
When ACLK_DISP clock frequency is calculated,
only pixel processing capacity is considered currently.
For the rotation case,
the short burst DMA request and line memory filling operation
take a long time to transmit the first data to DDI.
These are able to make the unwanted noise image when displaying
though the calculated ACLK_DISP is high enough to process pixel data.
So, to cope with this kind of issue, the ACLK_DISP calculating function
has been updated to consider the initial latency.
 <factors> default bus latency, rotating(exclusive with AFBC), scaling, DSC
It is expected that a higher ACLK_DISP will be determined to avoid problems
caused by the operation types.

Change-Id: I38657e44e8dea8f6398f3111289fd7782815c624
Signed-off-by: Jaehoe Yang <jaehoe.yang@samsung.com>