fbdev: dpu20: skip the flush workqueue of update handler
If update handler work is not remained when flush workqueue is called,
an update handler thread doesn't need to flush work in workqueue.

This patch detects remained work in workqueue and then if work is not
remained, flush workqueue is not called.

It helps to reduce time of hibernation entry.

Change-Id: I463dd5e3df9ff783b0295ce5378453e427dd005e
Signed-off-by: Jiun Yu <jiun.yu@samsung.com>