fbdev: dpu20: fix unspecified csc type handling error
When type of eq_mode is CSC_STANDARD_UNSPECIFIED,
driver handles it as CSC_BT_601.
Coefficient SFR setting mode(HARDWIRED/CUSTOMIZED) is based on csc type.
The condition code about SFR mode was initially determined to be CUSTOMIZED,
but it should be changed to HARDWIRED because csc type was changed to BT_601.
That is, Coef. SFR mode must be HARDWIRED for UNSPECIFIED type case.

Change-Id: Iefddf36a2605921540f3978af96d0c872d5028ba
Signed-off-by: Jaehoe Yang <jaehoe.yang@samsung.com>