media: smfc: wakeup waiter at compression timeout
exynos_smfc_wait_done() notifies the drivers of MC-scalers that
produces images to compress to a shared on-the-fly buffer. Since
signals between SMFC and MC-Scaler are implemented with a simple
counter and the synchronization between them are done by monitoring
the counter by SMFC, it is important not to reset the counter before
SMFC completes the current work.
However, MCScaler may waits for SMFC in exynos_smfc_wait_done()
indefinitely when compression in SMFC is timed out because the driver
does not wake up tasks blocked in exynos_smfc_wait_done() when the
timeout occurrs.

Change-Id: I3028eb6934ae92ca008d0eb0a98ef0471512aaf2
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>