fbdev: dpu20: change v-porch field write scheme
In EVT1, v-porch bit-field was extended to 16-bit.
The related changes were already applied.
This change is for value write scheme.
write_mask can't be used to write v-porch field in EVT1
because correct v-porch value can not be read from registers.
So, we changed the 32-bit value to write at once.
  (from) write value for each field    : dsim_write_mask
  (to)   write value for each register : dsim_write

Change-Id: Iac0fffeb4d96f66aff1db140f265cbd53ab921d9
Signed-off-by: Jaehoe Yang <jaehoe.yang@samsung.com>