[COMMON] media: mfc: expand encoder LAST_SEQ
Even if we don't have the src/dst buffer,
we can send the LAST_SEQ command.
And if the header only data(SPS or SPS+PPS not including IDR)
is created in dst buffer,
drvier update the tag with specific data("0xC5D")
becasue we don't have the proper tag.

Change-Id: If0a510b2793d65f9c2d2694fc1689436ffb8a7c8
Signed-off-by: Ayoung Sim <a.sim@samsung.com>