sched: ems: handle group migration for util-est
Util-est in EMS is applicable to each schedtune group. If a task moves from
group util-est=1 to group util-est=0, it should remove the .ewma or .enqueued
of task from .enqueued of cfs_rq. On the contrary, it should add .ewma or
.enqueued of task to .enqueued of cfs_rq.

Also, if group util-est become 0 to 1 or 1 to 0, .enqueued of cfs_rq to which
the tasks in the group belongs should be updated in the same manner as above.

Change-Id: I2e7a458649cc0acebfb98052eaae26177fab35a9
Signed-off-by: Park Bumgyu <bumgyu.park@samsung.com>