[9830] fbdev: dpu30: fix reconvey enable control
The recovery is enabled when buffer data is compressed to avoid stuck.
DPP driver sets AFBC and SBWC sequentially,
and recovery is enabled when each is enabled.
Because both AFBC and SBWC can't not be set at the same thime,
the enabled recovery setting in AFBC is disabled again
when SBWC related settings are made.

Change-Id: Icd4af314a845a30737489508cac681963d84949b
Signed-off-by: Jaehoe Yang <jaehoe.yang@samsung.com>