[9630] fbdev: dpu30: fix recovery enable control
The recovery is enabled when buffer data is compressed to avoid stuck.
DPP driver sets AFBC and SBWC sequentially,
and recovery is enabled when each is enabled.
Because both AFBC and SBWC can't not be set at the same thime,
the enabled recovery setting in AFBC is disabled again
when SBWC related settings are made.

Change-Id: I6ce85e94e1f084e92764921bdeae519d5415aec2
Signed-off-by: Kim Haksoo <herb@samsung.com>