[COMMON] soc: samsung: ppmpu: Do not handle TZASC interrupt in PPMPU interrupt handler
If TZASC and PPMPU share their interrupt line then
IRQ handlers are registered with IRQF_SHARED irqflag.
In this case, when the interrupt is generated, both
IRQ handlers are called and they will handle it.
If the interrupt is for TZASC then PPMPU IRQ handler
doesn't have to deal with it.

Therefore, this patch makes the PPMPU IRQ handler
pass TZASC interrupts without any other services.

Change-Id: Ie64a753f9690619570eb9c5f86e65fb9768f1d2c
Signed-off-by: Junho Choi <junhosj.choi@samsung.com>