[COMMON] soc: samsung: tzasc: Do not handle PPMPU interrupt in TZASC interrupt handler
If TZASC and PPMPU share their interrupt line then
IRQ handlers are registered with IRQF_SHARED irqflag.
In this case, when the interrupt is generated, both
IRQ handlers are called and they will handle it.
If the interrupt is for PPMPU then TZASC IRQ handler
doesn't have to deal with it.

Therefore, this patch makes the TZASC IRQ handler
pass PPMPU interrupts without any other services.

Change-Id: I3cb0b1f13b219f3f31162945e13462125193fd2a
Signed-off-by: Junho Choi <junhosj.choi@samsung.com>