fbdev: dpu30: Added for error processing for unsupported format.
When unsupported format is delivered, error handling is not performed properly, resulting in panic.
This is because the format information is taken before checking
whether the format received from the HWC is valid.
Therefore, the error checking was allowed when the format information was taken when the update area was set.
In the win_configuration process,
the position of decon_check_limitation was changed so that the format could be validated first.

Change-Id: Ia02113e3b3bd79fb444b03f769a1885d655b1814
Signed-off-by: yeongran shin <yr613.shin@samsung.com>