fbdev: dpu30: fix packet go error
If there is a situation in which you have to wait for fifo empty
in order to check the constraints when
you write command after dsim_write_cmd_set, fifo empty timeout may occur.
This is because it waits for empty without ready control
while packet go enabled.
Thus, packet go disable is added at the end of dsim_write_cmd_set.

Change-Id: I9b8255378cbc84f297a69f2ca0ab99354e144520
Signed-off-by: Jaehoe Yang <jaehoe.yang@samsung.com>