fbdev: dpu30: correct weak points in decon debugfs
Size from user is used without validation to allocate memory with kmalloc.
When kmalloc is called with size 0, it returns ZERO_SIZE_PTR,
which is equal (void*)16. The ptr is then passed to copy_from_user with the
same size, which if size is 0 does not check destination address and returns 0.
Finally, without any further checks, sscanf is called on the buffer,
which attempt to read it causing crash.

Change-Id: If493d0d3885484cc8887d1f267560e79963bd217
Signed-off-by: Kijun Kim <kijun.kim@samsung.com>