fbdev: dpu30: modify find_format to avoid NULL pointer dereference
Function `dpu_find_fmt_info()` returns pointer to the `dpu_fmt` structure
for which field `fmt` has the same value as the provided argument.
In case, structure with the requested value of `fmt` does not exist,
*NULL is returned*.
To avoid NULL pointer dereference in the caller,
return value is changed to default(ARGB8888) format instead of NULL.

Because it's called from many places,
if we make a change in all callers, it may be able to cause another error.
And, 'NULL ptr handling' or 'default value applying' can be
an undesirable display in either case,
So, the decision for the change was made in a simple way.

Change-Id: I4d2a46dd0ea728a9aaf115dfc77adf93fe4bdf98
Signed-off-by: Jaehoe Yang <jaehoe.yang@samsung.com>