iommu/samsung: add support for iova reserve
Iova areas need to be reserved for some reasons:
We can imagine a DMA device that is running since bootloader and needs
to keep running after kernel booting. It was running in the physical
address space before kernel booting and should seemlessly run after its
IOMMU is attached. For those devices, the DMA area the device is
accessing should be reserved from the bootloader and the iova area with
the identical address of the DMA area should be written on the page
table of the IOMMU when it is attached.
Another reason for the reserved iova area is for a DMA device that needs
some buffers should be in a specific area. One of the example is MFC.
The firmware address should be lower than all other buffers MFC should
access. To guarantee the restriction, MFC driver should reserve the DMA
address for the firmware erlier than dma address allocation for other
buffers.

"samsung,iommu-identity-map" property reserves and maps the specified
DMA area for the former reason and "samsung,iommu-reserved-map" only
reserves the specific area for the latter reason.

Those properties should be specified in the device tree node of the
client devices. The base address of a region should be #address-cells
wide and the size of the region should be #size-cells wide.

Here is an example with a device of two reserved identity mappings.

client {
    ...<snip>...
    samsung,iommu-identity-map = <0x0 0x10680000 0x10000>,
			         <0x0 0x10700000 0x10000>;
};

Change-Id: I37c5c26240a6588bc083290db4df8e4bae7fa790
Signed-off-by: Janghyuck Kim <janghyuck.kim@samsung.com>
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>