[FIXES] HACK: cal-if: Make it possible to build cal-if without PM enabled
[fix-min-config-build]

Changes:

 * Disable (with ifdefs) all PM-related code, so it's possible to build
   cal-if with PM disabled
 * cmucal.h: Move error codes up before vclk.h inclusion, as it uses one
   of those codes, which leads to build error (in case when PM disabled)
 * Add missing headers inclusion (when PM options are disabled, some
   cal-if headers won't be included, which leads to build errors)
 * Add CONFIG_EXYNOS_PM option (and disable it), so PM-related code can
   be disabled (otherwise it's enabled when CONFIG_ARCH_EXYNOS is
   defined)
 * cal-if.h: Remove check for CONFIG_PWRCAL, as that option doesn't
   occur in any Kconfig

Not 100% sure that all those changes are legit, so marking this patch as
a HACK for now. The main objective of this patch is to make it possible
to build the kernel with minimal support needed for serial console,
hence it's desired that Exynos PM is disabled, as it's quite huge. These
changes also make sense for further upstreaming, as it must be possible
to disable any option without breaking the build.

Signed-off-by: Sam Protsenko <semen.protsenko@linaro.org>
Change-Id: Id0b754c7ca697b697beabcf7125beb79f22921b9