fbdev: dpu30: fix vulnerability of user controllable file descriptor
DPU driver may get a wrong fd through the released file pointer,
if the user(HWC) closes the fd and releases the file pointer.
To avoid this possibility,
the position of fd_install was moved after work queue.

* existing code can cause problems at following point.
decon_create_release_fences(...)
 fd_install(rel_fence, get_file(sync_file->file)); <-- this point

For reference, this is similar to that implemented in the DRM/KMS framework

Change-Id: I3fc1da0a8c63a50439d344f2b45e64210b4d0272
Signed-off-by: Kijun Kim <kijun.kim@samsung.com>