dma-heap: samsung: Support protected heap
Register secure heaps, which have a protection ID defined by DT.
The secure device virtual address is mapped when allocated by smc.
If the device that has own iommu calls dma_buf_mat_attachment,
the secure device virtual address mapped in advance is used.

These heaps are used for DRM playback, etc.

The secure heaps are needed only when the system need to protect
secure contents. We add secure heaps if the content protection config
is enabled. If not, the trial of adding secure heap returns -ENODEV.
-ENODEV is not treated as error because these heaps are not needed.
However system heap treated as an error because the system heap is
always normal heap.

Change-Id: I90bb4e1eef2322c7e0c351ef52aaeca0da93ee47
Signed-off-by: Hyesoo Yu <hyesoo.yu@samsung.com>