[FIXES] fbdev: exynos: dpu30: Fail gracefully when there is no panel
Seems like dpu3.0 driver is written with idea that LCD panel *must* be
present on the board. Otherwise all sorts of issues can happen (not
tested code branches, kernel panics, intentional BUG() call, etc). Fix
DPU30 in a minimal manner, so that it's possible to run the kernel when
there is no panel present (e.g. on E850-96 board, where we only have
HDMI and not LCD panel).

Particularly this fixes next errors:
  - don't try to actually do dsim_read_data() if decon object wasn't
    initialized, to avoid NULL pointer dereference
  - don't free memory allocated with devm functions
  - wake up in kthread when kthread_stop() is issued
  - fail dsim_probe() if dsim_register_panel() failed
  - don't throw kernel panic in dsim_register_panel() if panel wasn't
    found, return error code instead

Signed-off-by: Sam Protsenko <semen.protsenko@linaro.org>
Change-Id: Ic0cfb9390cc67e84a3e017f96050d4aa0a5003d1