dma-heap: samsung: add property heap name on device tree
The device name was defined by reserved memory node or hard-coding
name. The hard-coding name is difficult to be managed. And if the user
wants to register cma, carveout heap both by using the same reserved memory,
the heap name could be duplicated and it would be an intialization race.
So, we change to define the heap name by device node explicitly.

In addition, system heap is probed by sw binding
to unify the heap registration with other heaps.

Change-Id: If985154daefd327d0f385b7b9e899c317f7efb3b
Signed-off-by: Hyesoo Yu <hyesoo.yu@samsung.com>