dma-heap: samsung: trace dmabuf reference
When out-of-memory happens and dmabuf leakage happens, we have
to show meaningful information to trace root cause of leakage.

If there is a leak in dmabuf, there can be three problems. The
first is the process's fd leak and the second is the process's
mmap leak, and the third is the device's reference leak.

For mmap, we register vm_operations to manage the relationship
between task and buffer. For fd, a list of tasks that accesses
dmabuf through syscall is stored and all fds of the task are
searched when the problem happens. For device, the relationship
between dmabuf and device is stored on samsung_dma_buffer.

On show_mem callbacks, we shows detailed buffer information
if it is dmabuf leak (for example, dmabuf total size is bigger
than 1/3 of the total managed memory, or total size is bigger
than 3GB.)

            comm   pid   fdrefcnt mmaprefcnt fdsize(kb) mmapsize(kb)
 composer@2.4-se   552         26         26     104736     104736
  surfaceflinger   636         40         40     110296     110296

  node                  exp       size   refcount mmaprefcnt devrefcnt attached device..
 36561      system-uncached     638976          3          1         1 18500000.mali(1)
 36562               system      40960          2          1         0

In addition to that, we always show the information of total
dma-heap size which is not managed by linux kernel to calculate
the free memory status on show_mem.

system: 20564 (kb)
system-uncached: 111332 (kb)

Change-Id: I365bd3092ef28490901080b81a14c5c4dfd8aefd
Signed-off-by: Hyesoo Yu <hyesoo.yu@samsung.com>