dma-heap: samsung: show attached device for hpa heap
The HPA buffers are referenced by attached device, and
file descriptor of process. These buffers don't need to
consider mmap count because secure buffer can not be mmaped.
The file descriptor might be tracked on dma-buf-trace
because that interates dma-buf file descriptor
when show_mem() calls, so we only show devices to attach
the hpa buffers.

Change-Id: Iafad9bf1bb9a51f440d8bf2e9c9c14cef1aa4d90
Signed-off-by: Hyesoo Yu <hyesoo.yu@samsung.com>