dma-heap: samsung: manage all devices when mapping and unmapping
dma-buf-trace manages the devices that is representative
device of the same iova domain. There are two problem.
First each device could have a different virtual address
due to dma-coherent even if they are in the same iova domain.
Second we don't know exactly which device in the same domain
is the problem.

We change it to manage the number of map and unmap for
every device. That helps to find reference leakage of
a specific device.

dma-buf-trace device's refcount was only incremented
on first iova map. The purpose is to calculate actual
buffer size referenced by device. For example, if a
device maps the same buffer twice, the buffer size
referenced by the device is counted once.

However there is logical problem. The dva is
shared amoung devices of the same iommu domain.
First referenced device may not be actual user
if we manage all devices. so we also remove that.

Change-Id: Ibfc5ae19def9e033a92585ed8e3e5ade0278cba3
Signed-off-by: Hyesoo Yu <hyesoo.yu@samsung.com>