wlbt: Fix for prevent issues
Changes are done to fix the prevent issues(RESOURCE LEAK, REVERSE_INULL)
in function slsi_set_vendor_ie and slsi_ioctl_fn_lookup.
Fixing the ioctl lookup function. It was giving out wrong index when
a substring of a command is provided and that is present before
the real function in the ioctl function table, so choosing
the wrong function to call.

SCSC-Bug-Id: HOST-14765
Change-Id: I5bc92c238d89be11160c78b5ffc42dc89aa71ddb
Signed-off-by: Sushant B Butta <b.sushant@samsung.com>