dma-heap: samsung: optimize to find the buffer and ref list
If hundreds or thousands of buffers are allocated, or if the
task has references to many buffers, it takes a lot of time
to find ref and buffer node.

We change for buffer to indicate the trace buffer node directly
to find the node O(1) and change the way to find refnode.
The refnode is linked both task node and buffer node.
We find the refnode to search the list of task's list.

However the number of task's refnode is bigger then the number
of buffer's refnode generally because one task might have reference
of hundreds or thousands of buffers but one buffer is unlikely
referenced by hundreds of thousands of tasks. It means finding
the buffer's refnode is more efficient, so we change the way
to find refnode.

Change-Id: I67f3000b787be4715e11161ecb255a347682da79
Signed-off-by: Hyesoo Yu <hyesoo.yu@samsung.com>