[FIXES] exynos-coresight: Fix kernel panic when trying to read PCSR
PCSR reading procedure is probably implemented incorrectly, which leads
to kernel panic. When lockup happens, Exynos CoreSight driver tries to
read PCSR register, but that leads to next error:

    Synchronous External Abort: synchronous external abort (0x96000010)

which happens in exynos_cs_get_pc() and leads further to kernel panic.
I'm not sure how to fix that properly, so just disable that for now,
which is enough to overcome the kernel panic condition.

Signed-off-by: Sam Protsenko <semen.protsenko@linaro.org>