media: scaler: fix the bugs about sc_remove()
There were these problems.

1. the clocks were released after released.
- clocks was got by function of devm. So, these will be released by devres automatically.
But, the driver was releasing manually with using the function of not devm.
It makes releasing occur twice.
So, remove the function-call about releasing the clocks.

2. the device node was not removed.
- video_unregister_device() was not called in sc_remove().
So, add function-call of this.

3. the misc-device of scaler-ext was not removed.
In sc_remove(), the device of scaler-ext was not removed.
And it makes the list of character devices broken down.
So, add function-call to remove this.

Change-Id: Ia179cefdcf92fbf8739e4eb26d0ee751875df32f
Signed-off-by: Dongseong Kim <ds_in_ds.kim@samsung.com>