media: scaler: remove calling func to reserve memory for vOTF
For vOTF, the driver should map the memory by 1:1 and reserve the memory.
In past, the driver used iommu_map() and iommu_dma_reserve_iova() for that.
But, it is not supported anymore.
And now, the driver does that by writing the information about that into DT.
So, remove these function-call.

Change-Id: I7f21694a33f2bb158003c76139163d111ea1d577
Signed-off-by: Dongseong Kim <ds_in_ds.kim@samsung.com>