media: smfc: update code for minlock and bts
SMFC of some product is belonging to not INT but M2M.
So, to improve the performance by requesting DVFS,
the driver of them should request it to M2M not INT.
However, INT minlock is still required for bus performance
and MIF minlock is also required. For supporting this,
SMFC driver set minlock of M2M/INT/MIF and updates
a bandwidth to BTS driver to add a minlock for INT and MIF.

Information for minlock and bts should be updated into DT
and the driver do parsing and use it.

Change-Id: Iabea04a15e6e9775c651a16b90e39a38ad577b24
Signed-off-by: Seungchul Kim <sc377.kim@samsung.com>