Revert "ANDROID: Sched: Add restricted vendor hooks for scheduler"
This reverts most of commit 292f430816f36384634ca1365ccc67c08302dd21.

Only a small number of the original hooks in this commit were ever used,
so remove all of the unused ones.  The hooks removed are:
	android_rvh_try_to_wake_up
	android_rvh_try_to_wake_up_success
	android_rvh_wake_up_new_task
	android_rvh_new_task_stats
	android_rvh_flush_task
	android_rvh_tick_entry
	android_rvh_schedule
	android_rvh_sched_cpu_starting
	android_rvh_sched_cpu_dying
	android_rvh_account_irq
	android_rvh_place_entity
	android_rvh_update_cpu_capacity
	android_rvh_update_misfit_status

If these are needed by any real user, it can easily be reverted to add it
back and then the symbols must be added to the abi list at the same time
to prevent it from being removed again later.

Bug: 203756332
Bug: 173725277
Cc: Shaleen Agrawal <shalagra@codeaurora.org>
Cc: Pavankumar Kondeti <pkondeti@codeaurora.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>
Change-Id: Icc24ed128fe1034ad663b749caa3bc2e2d0efac0