[COMMON] usb: dwc3: Check suspend status in pullup
Even if USB cable is not connected, gadget run and irq is enabled
through get sync in pullup. This causes a problem that dwc->connected
becomes true due to reset interrupt, and dwc3_runtime_suspend/idle fail.
Therefore, always check the suspend status in pullup and return if it is
in suspend state.

Change-Id: Ib2c0467b7f434bdd1c6a38f106ef2c4b0f6e09c0
Signed-off-by: Oh Eomji <eomji.oh@samsung.com>