[COMMON] usb: gadget: modify gadget->connected location
If gadget->connected is false in dwc3_gadget_resume, the gadget start is
skipped. gadget->connected becomes true after pullup is completed.
Therefore, if the resume is performed before pullup is completed in the
usb connection sequence, the gadget start is skipped. So we set the
gadget->connected true before performing pullup.

Change-Id: Ib77cc7f03625291b0e47b724aac986d67dca55cb
Signed-off-by: Oh Eomji <eomji.oh@samsung.com>