soc: samsung/debug: dss: change judging secondary cpu lockup
The core lockup cannot be determined using the existing
method by removing the cpu from the online cpu mask from
the arm64 nmi panic handler. To solve this problem, the
dss context saved cpu mask is newly defined and the lockup
is determined by comparing it with the cpu active mask.

Change-Id: I537f461d7f445b15c98814626210d770fee565e1
Signed-off-by: Donghyeok Choe <d7271.choe@samsung.com>