wlbt: Send scan result to cfg80211 in sync_ind
This patch is to send scan result to cfg80211 in sync_ind.
Scan result(IEs) is used in wpa_supplicant when exteranl
auth case in order to check SAE H2E of AP.
However if there is no BSS which device try to connect,
wpa_supplciant cann't set H2E bit, so device cann't connect
to AP. So scan result is sent in case of external auth
in sync_ind.

SCSC-Bug-Id: HOST-16071

Change-Id: I1e4eabcc1137dc35c739f13644f7521f464b1617
Signed-off-by: Shubha C <shubha.c@samsung.com>