phy: exynos5-usbdrd: Add 26MHz ref clk support
26MHz reference clock is used on modern Exynos chips (like Exynos850).
And although for some PHY controllers (like one found on Exynos850)
there is no need to set the refclk frequency (and corresponding bits in
CLKRSTCTRL[7:5] are marked as RESERVED), its support still has to be
added in the driver. Otherwise exynos5_rate_to_clk() fails, which leads
in turn to probe error.

Add the correct value for 26MHz refclk to make it possible to add
support for new Exynos USB DRD PHY controllers.

Signed-off-by: Sam Protsenko <semen.protsenko@linaro.org>